backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0

sqlalchemy>=2.0.29
alembic>=1.13.1
psycopg2-binary>=2.9.9
pydantic>=2.6.4
pydantic[email]
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

redis>=5.0.3
httpx>=0.24.0

python-dotenv
asyncpg
aiofiles
aiosmtplib
python-multipart